+++ src/anomalib/data/__init__.py
@@ -48,7 +48,7 @@ from .dataclasses import (
 
 # Datamodules
 from .datamodules.base import AnomalibDataModule
-from .datamodules.depth import DepthDataFormat, Folder3D, MVTec3D
+from .datamodules.depth import ADAM3D, DepthDataFormat, Folder3D, MVTec3D
 from .datamodules.image import (
     BMAD,
     MPDD,
@@ -70,7 +70,7 @@ from .datamodules.video import Avenue, ShanghaiTech, UCSDped, VideoDataFormat
 
 # Datasets
 from .datasets import AnomalibDataset
-from .datasets.depth import Folder3DDataset, MVTec3DDataset
+from .datasets.depth import ADAM3DDataset, Folder3DDataset, MVTec3DDataset
 from .datasets.image import (
     BMADDataset,
     BTechDataset,
@@ -177,6 +177,7 @@ __all__ = [
     # Depth Data Modules
     "Folder3D",
     "MVTec3D",
+    "ADAM3D",
     # Image Data Modules
     "BMAD",
     "BTech",
@@ -204,6 +205,7 @@ __all__ = [
     "FolderDataset",
     "KolektorDataset",
     "MPDDDataset",
+    "ADAM3DDataset",
     "MVTecADDataset",
     "MVTecLOCODataset",
     "TabularDataset",