+++ src/anomalib/data/datamodules/depth/__init__.py
@@ -5,6 +5,7 @@
 
 from enum import Enum
 
+from .adam_3d import ADAM3D
 from .folder_3d import Folder3D
 from .mvtec_3d import MVTec3D
 
@@ -14,6 +15,7 @@ class DepthDataFormat(str, Enum):
 
     MVTEC_3D = "mvtec_3d"
     FOLDER_3D = "folder_3d"
+    ADAM_3D = "adam_3d"
 
 
-__all__ = ["Folder3D", "MVTec3D"]
+__all__ = ["Folder3D", "MVTec3D", "ADAM3D"]